fix some gcc 4.9 valid warnings
Scilab types used an int size whereas some natives functions return a
size_t. static_cast<int> is needed.

`for(iterator i = begin(), end = cases->end(); i != end; ++i)` will
disable optimizations as the end variable is not `const` thus might be
modified in the loop.

+++ scilab/modules/boolean/sci_gateway/cpp/sci_bool2s.cpp
@@ -82,7 +82,7 @@ types::Function::ReturnValue sci_bool2s(types::typed_list &in, int _iRetCount, t
 
         types::Sparse* pSpOut = new types::Sparse(pSpIn->getRows(), pSpIn->getCols());
 
-        size_t iNonZeros = pSpIn->nonZeros();
+        int iNonZeros = static_cast<int>(pSpIn->nonZeros());
 
         //coords
         int* pRows = new int[iNonZeros * 2];
@@ -106,7 +106,7 @@ types::Function::ReturnValue sci_bool2s(types::typed_list &in, int _iRetCount, t
         types::SparseBool* pSpbIn = in[0]->getAs<types::SparseBool>();
         types::Sparse* pSpOut = new types::Sparse(pSpbIn ->getRows(), pSpbIn ->getCols());
 
-        size_t iNonZeros = pSpbIn ->nbTrue();
+        int iNonZeros = static_cast<int>(pSpbIn->nbTrue());
 
         //coords
         int* pRows = new int[iNonZeros * 2];
